star: make the ram_console not depend on the carveout size
Currently, the carveout size is fixed as 128M
and the ram_console is fixed and depend on the CarveoutSize.

this patch make the ram_console and the *temporary* reserved_buffer
not depend on the carveout size.

and the commit "[HACK] copy the warmboot information to the original
reserved_buffer area before shutdown" is needed to fix reboot issue.