mm.h: use separate shrinker.h
A callback you can register to apply pressure to ageable caches.
'sc' is passed shrink_control which includes a count 'nr_to_scan'
and a 'gfpmask'.  It should look through the least-recently-used
'nr_to_scan' entries and attempt to free them up.  It should return
the number of objects which remain in the cache.  If it returns -1, it means
it cannot do any scanning at this time (eg. there is a risk of deadlock).
The callback must not return -1 if nr_to_scan is zero.

The 'gfpmask' refers to the allocation we are currently trying to
fulfil.

Note that 'shrink' will be passed nr_to_scan == 0 when the VM is
querying the cache size, so a fastpath for that case is appropriate.