OTF workaround: AVP, GPU, VDE freq adjusteable from the cmdline.
This is the only way, how can be this values adjusted.
The sysfs interface remain only for information.

Usage example (put this to the end of the cmdline)
avpfreq=240000 gpufreq=333000 vdefreq=650000

The values is limited, so if you used a value within the acceptable range
then the default value will be set.

Possible range:
AVP: 200000-280000
GPU: 300000-366000
VDE: 600000-700000